backend/eprocurement_portal/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.2
python-decouple>=3.7
pillow>=10.0.0